+++ mysql-update/0.1.3.sql
@@ -1,0 +1,2 @@
+ALTER TABLE `mailChain` ADD `sentDirect` BOOLEAN NOT NULL DEFAULT FALSE AFTER `removeAuthenticatedSender`;
+ALTER TABLE `mailLog` ADD `outgoingID` INT(255) NOT NULL AFTER `sendDate`, ADD INDEX (`outgoingID`);